+++ sdk/ml/azure-ai-ml/tests/internal_utils/unittests/test_ml_client.py
@@ -12,6 +12,7 @@ from azure.ai.ml import (
     load_batch_endpoint,
     load_component,
     load_compute,
+    load_connection,
     load_data,
     load_datastore,
     load_environment,
@@ -21,12 +22,10 @@ from azure.ai.ml import (
     load_online_endpoint,
     load_registry,
     load_workspace,
-    load_connection,
 )
 from azure.ai.ml._azure_environments import AzureEnvironments, _get_default_cloud_name
 from azure.ai.ml._scope_dependent_operations import OperationScope
-from azure.ai.ml._telemetry import get_appinsights_log_handler
-from azure.ai.ml._telemetry.logging_handler import AzureMLSDKLogHandler
+from azure.ai.ml._telemetry import set_appinsights_distro
 from azure.ai.ml._user_agent import USER_AGENT
 from azure.ai.ml.constants._common import AZUREML_CLOUD_ENV_NAME
 from azure.ai.ml.exceptions import ValidationException
@@ -531,7 +530,8 @@ class TestMachineLearningClient:
         assert ml_client._base_url == "https://test.management.azure.com"
         assert _get_default_cloud_name() == "test_cloud"
 
-    def test_enable_telemetry(self) -> None:
+    @patch("azure.ai.ml._telemetry.logging_handler.configure_azure_monitor")
+    def test_enable_telemetry(self, mock_configure_azure_monitor) -> None:
         subscription_id = "fake-sub-id"
         resource_group_name = "fake-rg-name"
 
@@ -553,11 +553,9 @@ class TestMachineLearningClient:
                 "subscription_id": subscription_id,
                 "resource_group_name": resource_group_name,
             }
-            handler, _ = get_appinsights_log_handler(
-                USER_AGENT, **{"properties": properties}, enable_telemetry=enable_telemetry
-            )
+            set_appinsights_distro(USER_AGENT, **{"properties": properties}, enable_telemetry=enable_telemetry)
             assert enable_telemetry
-            assert isinstance(handler, logging.NullHandler)
+            mock_configure_azure_monitor.assert_not_called()
 
         # confirm that telemetry is ENABLED when in jupyter notebook and enable_telemetry=True
         with patch("azure.ai.ml._telemetry.logging_handler.in_jupyter_notebook", return_value=True):
@@ -565,11 +563,10 @@ class TestMachineLearningClient:
                 "subscription_id": subscription_id,
                 "resource_group_name": resource_group_name,
             }
-            handler, _ = get_appinsights_log_handler(
-                USER_AGENT, **{"properties": properties}, enable_telemetry=enable_telemetry
-            )
+            set_appinsights_distro(USER_AGENT, **{"properties": properties}, enable_telemetry=enable_telemetry)
             assert enable_telemetry
-            assert isinstance(handler, AzureMLSDKLogHandler)
+            mock_configure_azure_monitor.assert_called_once()
+            mock_configure_azure_monitor.reset_mock()
 
         client = MLClient(
             credential=DefaultAzureCredential(),
@@ -590,11 +587,9 @@ class TestMachineLearningClient:
                 "subscription_id": subscription_id,
                 "resource_group_name": resource_group_name,
             }
-            handler, _ = get_appinsights_log_handler(
-                USER_AGENT, **{"properties": properties}, enable_telemetry=enable_telemetry
-            )
+            set_appinsights_distro(USER_AGENT, **{"properties": properties}, enable_telemetry=enable_telemetry)
             assert not enable_telemetry
-            assert isinstance(handler, logging.NullHandler)
+            mock_configure_azure_monitor.assert_not_called()
 
         # confirm that telemetry is DISABLED when in jupyter notebook and enable_telemetry=False
         with patch("azure.ai.ml._telemetry.logging_handler.in_jupyter_notebook", return_value=True):
@@ -602,8 +597,6 @@ class TestMachineLearningClient:
                 "subscription_id": subscription_id,
                 "resource_group_name": resource_group_name,
             }
-            handler, _ = get_appinsights_log_handler(
-                USER_AGENT, **{"properties": properties}, enable_telemetry=enable_telemetry
-            )
+            set_appinsights_distro(USER_AGENT, **{"properties": properties}, enable_telemetry=enable_telemetry)
             assert not enable_telemetry
-            assert isinstance(handler, logging.NullHandler)
+            mock_configure_azure_monitor.assert_not_called()